requirements.txt
Django>=1.11,<2.0
Pillow==4.2.1
psycopg2==2.7.3.1
djangorestframework==3.7.1
